+++ src/main/java/com/ccsw/bidoffice/sector/model/SectorEntity.java
@@ -22,7 +22,7 @@ public class SectorEntity {
     private String name;
 
     @Column(name = "priority", nullable = false)
-    private Integer priority;
+    private Long priority;
 
     @Column(name = "start_date")
     private LocalDate startDate;
@@ -46,11 +46,11 @@ public class SectorEntity {
         this.name = name;
     }
 
-    public Integer getPriority() {
+    public Long getPriority() {
         return priority;
     }
 
-    public void setPriority(Integer priority) {
+    public void setPriority(Long priority) {
         this.priority = priority;
     }
